update kernel
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@58657 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/universal/linux-4.14/include/linux/blkdev.h
@@ -57,7 +57,7 @@ struct blk_stat_callback;
  */
 #define BLKCG_MAX_POLS		3
 
-static inline int blk_validate_block_size(unsigned int bsize)
+static inline int blk_validate_block_size(unsigned long bsize)
 {
 	if (bsize < 512 || bsize > PAGE_SIZE || !is_power_of_2(bsize))
 		return -EINVAL;
